Test fixes and some refactoring
