+++ cypress/integration/WebInterface/Test Cases/Validations/TestCaseRunTestCaseButtonValidations.spec.ts
@@ -1,0 +1,619 @@
+import {OktaLogin} from "../../../../Shared/OktaLogin"
+import {CreateMeasurePage} from "../../../../Shared/CreateMeasurePage"
+import {MeasuresPage} from "../../../../Shared/MeasuresPage"
+import {TestCasesPage} from "../../../../Shared/TestCasesPage"
+import {EditMeasurePage} from "../../../../Shared/EditMeasurePage"
+import {TestCaseJson} from "../../../../Shared/TestCaseJson"
+import {Utilities} from "../../../../Shared/Utilities"
+import {MeasureGroupPage} from "../../../../Shared/MeasureGroupPage"
+
+let measureName = 'TestMeasure' + Date.now()
+let CqlLibraryName = 'TestLibrary' + Date.now()
+let measurePath = 'cypress/fixtures/measureId'
+let measureGroupPath = 'cypress/fixtures/groupId'
+let PopIniPop = 'ipp'
+let PopNum = 'denom'
+let PopDenom = 'num'
+let PopDenex = 'ipp'
+let PopDenexcep = 'denom'
+let PopNumex = 'num'
+let testCaseTitle = 'test case title'
+let testCaseDescription = 'DENOMFail' + Date.now()
+let validTestCaseJson = TestCaseJson.TestCaseJson_Valid
+let invalidTestCaseJson = TestCaseJson.TestCaseJson_Invalid
+let testCaseSeries = 'SBTestSeries'
+
+describe('Test Case Run Test Case button validations', () => {
+
+
+    beforeEach('Login and Create Measure', () => {
+        CqlLibraryName = 'TestLibrary2' + Date.now()
+        //Create New Measure
+        CreateMeasurePage.CreateQICoreMeasureAPI(measureName, CqlLibraryName)
+        cy.setAccessTokenCookie()
+        cy.getCookie('accessToken').then((accessToken) => {
+            cy.readFile(measurePath).should('exist').then((fileContents) => {
+                cy.request({
+                    url: '/api/measures/' + fileContents + '/groups/',
+                    method: 'POST',
+                    headers: {
+                        authorization: 'Bearer ' + accessToken.value
+                    },
+                    // body: {
+                    //     "scoring": 'Proportion',
+                    //     "population":
+                    //     {
+                    //         "initialPopulation": PopIniPop,
+                    //         "denominator": PopDenom,
+                    //         "denominatorExclusion": PopDenex,
+                    //         "denominatorException": PopDenexcep,
+                    //         "numerator": PopNum,
+                    //         "numeratorExclusion": PopNumex
+                    //     }
+                    // }
+                    body: {
+                        "scoring": 'Proportion',
+                        "populations": [
+                            {
+                                "name": "initialPopulation",
+                                "definition": PopIniPop
+                            },
+                            {
+                                "name": "denominator",
+                                "definition": PopDenom
+                            },
+                            {
+                                "name": "denominatorExclusion",
+                                "definition": PopDenex
+                            },
+                            {
+                                "name": "denominatorException",
+                                "definition": PopDenexcep
+                            },
+                            {
+                                "name": "numerator",
+                                "definition": PopNum
+                            },
+                            {
+                                "name": "numeratorExclusion",
+                                "definition": PopNumex
+                            }
+                        ],
+                        "measureGroupTypes": [
+                            "Outcome"
+                        ]
+                    }
+                }).then((response) => {
+                        expect(response.status).to.eql(201)
+                        expect(response.body.id).to.be.exist
+                        cy.writeFile(measureGroupPath, response.body.id)
+                })
+            })
+        })
+        OktaLogin.Login()
+
+
+    })
+    afterEach('Logout and Clean up', () => {
+        OktaLogin.Logout()
+        Utilities.deleteMeasure(measureName, CqlLibraryName)
+
+    })
+    it('Run Test Case button is disabled  -- CQL Errors', () => {
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Add CQL
+        cy.get(EditMeasurePage.cqlEditorTab).click()
+
+        cy.readFile('cypress/fixtures/CQLForTestCaseExecution.txt').should('exist').then((fileContents) => {
+            cy.get(EditMeasurePage.cqlEditorTextBox).type(fileContents)
+        })
+
+        cy.get(EditMeasurePage.cqlEditorTextBox).type('{home}')
+        cy.get(EditMeasurePage.cqlEditorTextBox).type('adjfajsdfljafja;lsdjf')
+        
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.enabled')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).click()
+        
+        //Create Measure Group
+        cy.get(EditMeasurePage.measureGroupsTab).should('be.visible')
+        cy.get(EditMeasurePage.measureGroupsTab).click()
+        
+        cy.get(MeasureGroupPage.initialPopulationSelect).select('ipp')
+        cy.get(MeasureGroupPage.denominatorSelect).select('denom')
+        cy.get(MeasureGroupPage.numeratorSelect).select('num')
+        cy.get(MeasureGroupPage.numeratorExclusionSelect).select('num')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.visible')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.enabled')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).click()
+        
+        //validation successful save message
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('exist')
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('contain.text', 'Population details for this group updated successfully.')
+
+        TestCasesPage.createTestCase(testCaseTitle, testCaseDescription, testCaseSeries, validTestCaseJson, true)
+
+        TestCasesPage.clickEditforCreatedTestCase()
+        cy.get(TestCasesPage.runTestButton).should('be.disabled')
+
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+
+        cy.get(TestCasesPage.executeTestCaseButton).should('be.disabled')
+
+    })
+
+    it('Run / Execute Test Case button is disabled  -- Missing group / population selections', () => {
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Add CQL
+        cy.get(EditMeasurePage.cqlEditorTab).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorTab).click()
+
+        cy.readFile('cypress/fixtures/CQLForTestCaseExecution.txt').should('exist').then((fileContents) => {
+            cy.get(EditMeasurePage.cqlEditorTextBox).type(fileContents)
+        })
+        
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.enabled')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).click()
+        //cy.get(CQLEditorPage.successfulCQLSaveNoErrors).should('be.visible')
+
+        cy.get(EditMeasurePage.measureGroupsTab).should('be.visible')
+        cy.get(EditMeasurePage.measureGroupsTab).click()
+        OktaLogin.Logout()
+        cy.setAccessTokenCookie()
+        cy.getCookie('accessToken').then((accessToken) => {
+            cy.readFile('cypress/fixtures/measureId').should('exist').then((fileContents) => {
+                cy.request({
+                    url: '/api/measures/' + fileContents + '/groups/',
+                    method: 'PUT',
+                    headers: {
+                        authorization: 'Bearer ' + accessToken.value
+                    },
+                    body: {
+                        "scoring": 'Proportion',
+                        "populations": [
+                            {
+                                "name": "initialPopulation",
+                                "definition": PopIniPop
+                            },
+                            {
+                                "name": "denominator",
+                                "definition": PopDenom
+                            },
+                            {
+                                "name": "denominatorExclusion",
+                                "definition": PopDenex
+                            },
+                            {
+                                "name": "denominatorException",
+                                "definition": PopDenexcep
+                            },
+                            {
+                                "name": "numerator",
+                                "definition": PopNum
+                            },
+                            {
+                                "name": "numeratorExclusion",
+                                "definition": PopNumex
+                            }
+                        ],
+                        "measureGroupTypes": [
+                            "Outcome"
+                        ]
+                    }
+                }).then((response) => {
+                    expect(response.status).to.eql(200)
+                    expect(response.body.id).to.be.exist
+
+                })
+            })
+        })
+        OktaLogin.Login()
+        cy.wait(10000)
+
+        TestCasesPage.createTestCase(testCaseTitle, testCaseDescription, testCaseSeries, validTestCaseJson, true)
+
+        TestCasesPage.clickEditforCreatedTestCase()
+        cy.get(TestCasesPage.runTestButton).should('be.visible')
+        cy.get(TestCasesPage.runTestButton).should('be.disabled')
+
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+
+        cy.get(TestCasesPage.executeTestCaseButton).should('be.disabled')        
+    })
+    it('Run Test Case button is disabled -- Invalid TC Json', () =>{
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Add CQL
+        cy.get(EditMeasurePage.cqlEditorTab).click()
+        
+        cy.readFile('cypress/fixtures/CQLForTestCaseExecution.txt').should('exist').then((fileContents) => {
+            cy.get(EditMeasurePage.cqlEditorTextBox).type(fileContents)
+        })
+                
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.enabled')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).click()
+
+                
+        //Create Measure Group
+        cy.get(EditMeasurePage.measureGroupsTab).should('be.visible')
+        cy.get(EditMeasurePage.measureGroupsTab).click()
+                
+        cy.get(MeasureGroupPage.initialPopulationSelect).select('ipp')
+        cy.get(MeasureGroupPage.denominatorSelect).select('denom')
+        cy.get(MeasureGroupPage.numeratorSelect).select('num')
+        cy.get(MeasureGroupPage.numeratorExclusionSelect).select('num')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.visible')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.enabled')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).click()
+                
+        //validation successful save message
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('exist')
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('contain.text', 'Population details for this group updated successfully.')
+        
+        //Navigate to Test Cases page and add Test Case details
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+        cy.get(TestCasesPage.newTestCaseButton).should('be.visible')
+        cy.get(TestCasesPage.newTestCaseButton).should('be.enabled')
+        cy.get(TestCasesPage.newTestCaseButton).click()
+
+        cy.get(TestCasesPage.detailsTab).click()
+        
+        cy.get(TestCasesPage.testCasePopulationList).should('be.visible')
+        
+        cy.get(TestCasesPage.testCaseTitle).should('be.visible')
+        cy.get(TestCasesPage.testCaseTitle).should('be.enabled')
+        cy.get(TestCasesPage.testCaseTitle).type(testCaseTitle, { force: true })
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).type(testCaseDescription)
+        cy.get(TestCasesPage.testCaseSeriesTextBox).type(testCaseSeries).type('{enter}')
+        
+        //Add json to the test case
+        cy.get(TestCasesPage.aceEditor).type(invalidTestCaseJson)
+        
+        cy.readFile('cypress/fixtures/measureId').should('exist').then((id)=> {
+            cy.intercept('POST', '/api/measures/' + id + '/test-cases').as('testcase')
+        
+            cy.get(TestCasesPage.createTestCaseButton).should('be.visible')
+            cy.get(TestCasesPage.createTestCaseButton).should('be.enabled')
+            cy.get(TestCasesPage.createTestCaseButton).click()
+        
+            //saving testCaseId to file to use later
+            cy.wait('@testcase').then(({response}) => {
+                expect(response.statusCode).to.eq(201)
+                cy.writeFile('cypress/fixtures/testCaseId', response.body.id)
+            })
+
+            cy.get(TestCasesPage.confirmationMsg).should('have.text', 'An error occurred with the Test Case JSON while creating the test case')
+            cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+            cy.get(EditMeasurePage.testCasesTab).click()
+        
+        })
+        
+        TestCasesPage.clickEditforCreatedTestCase()
+        cy.get(TestCasesPage.runTestButton).should('be.disabled')
+        
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+        
+    })
+    //MAT-4421  -- the execute button is enabled when it shouldn't be
+    it.skip('Execute Test Case button is disabled  -- Invalid TC Json', () => {
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Add CQL
+        cy.get(EditMeasurePage.cqlEditorTab).click()
+        
+        cy.readFile('cypress/fixtures/CQLForTestCaseExecution.txt').should('exist').then((fileContents) => {
+            cy.get(EditMeasurePage.cqlEditorTextBox).type(fileContents)
+        })
+                
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.enabled')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).click()
+
+                
+        //Create Measure Group
+        cy.get(EditMeasurePage.measureGroupsTab).should('be.visible')
+        cy.get(EditMeasurePage.measureGroupsTab).click()
+
+        //select a group type
+        cy.get(MeasureGroupPage.measureGroupTypeSelect).should('exist')
+        cy.get(MeasureGroupPage.measureGroupTypeSelect).should('be.visible')
+        cy.get(MeasureGroupPage.measureGroupTypeSelect).click()
+        cy.get(MeasureGroupPage.measureGroupTypeCheckbox).each(($ele) => {
+            if ($ele.text() == "Process") {
+                cy.wrap($ele).should('exist')
+                cy.wrap($ele).focus()
+                cy.wrap($ele).click()
+            }
+        })
+        cy.get(MeasureGroupPage.measureGroupTypeDropdownBtn).should('exist').invoke('click')        
+                
+        cy.get(MeasureGroupPage.initialPopulationSelect).select('ipp')
+        cy.get(MeasureGroupPage.denominatorSelect).select('denom')
+        cy.get(MeasureGroupPage.numeratorSelect).select('num')
+        cy.get(MeasureGroupPage.numeratorExclusionSelect).select('num')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.visible')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.enabled')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).click()
+                
+        //validation successful save message
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('exist')
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('contain.text', 'Population details for this group saved successfully.')
+        
+        //Navigate to Test Cases page and add Test Case details
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+        cy.get(TestCasesPage.newTestCaseButton).should('be.visible')
+        cy.get(TestCasesPage.newTestCaseButton).should('be.enabled')
+        cy.get(TestCasesPage.newTestCaseButton).click()
+        
+        cy.get(TestCasesPage.testCasePopulationList).should('be.visible')
+        
+        cy.get(TestCasesPage.testCaseTitle).should('be.visible')
+        cy.get(TestCasesPage.testCaseTitle).should('be.enabled')
+        cy.get(TestCasesPage.testCaseTitle).type(testCaseTitle, { force: true })
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).type(testCaseDescription)
+        cy.get(TestCasesPage.testCaseSeriesTextBox).type(testCaseSeries).type('{enter}')
+        
+        //Add json to the test case
+        cy.get(TestCasesPage.aceEditor).type(invalidTestCaseJson)
+        
+        cy.readFile('cypress/fixtures/measureId').should('exist').then((id)=> {
+            cy.intercept('POST', '/api/measures/' + id + '/test-cases').as('testcase')
+        
+            cy.get(TestCasesPage.createTestCaseButton).should('be.visible')
+            cy.get(TestCasesPage.createTestCaseButton).should('be.enabled')
+            cy.get(TestCasesPage.createTestCaseButton).click()
+        
+            //saving testCaseId to file to use later
+            cy.wait('@testcase').then(({response}) => {
+                expect(response.statusCode).to.eq(201)
+                cy.writeFile('cypress/fixtures/testCaseId', response.body.id)
+            })
+
+            cy.get(TestCasesPage.confirmationMsg).should('have.text', 'An error occurred with the Test Case JSON while creating the test case')
+            cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+            cy.get(EditMeasurePage.testCasesTab).click()
+        
+        })
+        
+        TestCasesPage.clickEditforCreatedTestCase()
+        cy.get(TestCasesPage.runTestButton).should('be.disabled')
+        
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+        
+        cy.get(TestCasesPage.executeTestCaseButton).should('be.disabled')
+        
+    })
+    it('Run Test Case button is disabled -- missing TC Json',() =>{
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Add CQL
+        cy.get(EditMeasurePage.cqlEditorTab).click()
+
+        cy.readFile('cypress/fixtures/CQLForTestCaseExecution.txt').should('exist').then((fileContents) => {
+            cy.get(EditMeasurePage.cqlEditorTextBox).type(fileContents)
+        })
+
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.enabled')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).click()
+        
+        //Create Measure Group
+        cy.get(EditMeasurePage.measureGroupsTab).should('be.visible')
+        cy.get(EditMeasurePage.measureGroupsTab).click()
+        
+        cy.get(MeasureGroupPage.initialPopulationSelect).select('ipp')
+        cy.get(MeasureGroupPage.denominatorSelect).select('denom')
+        cy.get(MeasureGroupPage.numeratorSelect).select('num')
+        cy.get(MeasureGroupPage.numeratorExclusionSelect).select('num')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.visible')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.enabled')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).click()
+        
+        //validation successful save message
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('exist')
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('contain.text', 'Population details for this group updated successfully.')
+
+        //Navigate to Test Cases page and add Test Case details
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+        cy.get(TestCasesPage.newTestCaseButton).should('be.visible')
+        cy.get(TestCasesPage.newTestCaseButton).should('be.enabled')
+        cy.get(TestCasesPage.newTestCaseButton).click()
+
+        cy.get(TestCasesPage.detailsTab).click()
+        
+        cy.get(TestCasesPage.testCasePopulationList).should('be.visible')
+        
+        cy.get(TestCasesPage.testCaseTitle).should('be.visible')
+        cy.get(TestCasesPage.testCaseTitle).should('be.enabled')
+        cy.get(TestCasesPage.testCaseTitle).type(testCaseTitle, { force: true })
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).type(testCaseDescription)
+        cy.get(TestCasesPage.testCaseSeriesTextBox).type(testCaseSeries).type('{enter}')
+
+        cy.readFile('cypress/fixtures/measureId').should('exist').then((id)=> {
+            cy.intercept('POST', '/api/measures/' + id + '/test-cases').as('testcase')
+        
+            cy.get(TestCasesPage.createTestCaseButton).should('be.visible')
+            cy.get(TestCasesPage.createTestCaseButton).should('be.enabled')
+            cy.get(TestCasesPage.createTestCaseButton).click()
+        
+            //saving testCaseId to file to use later
+            cy.wait('@testcase').then(({response}) => {
+                expect(response.statusCode).to.eq(201)
+                cy.writeFile('cypress/fixtures/testCaseId', response.body.id)
+            })
+
+            cy.get(TestCasesPage.confirmationMsg).should('contains.text', 'Test case created successfully!')
+            
+            cy.get(EditMeasurePage.testCasesTab).click()
+        
+        })
+
+        TestCasesPage.clickEditforCreatedTestCase()
+        cy.get(TestCasesPage.runTestButton).should('be.disabled')
+
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+
+    })
+    //MAT-4421  -- the execute button is enabled when it shouldn't be
+    it.skip('Execute Test Case button is disabled  -- missing TC Json', () => {
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Add CQL
+        cy.get(EditMeasurePage.cqlEditorTab).click()
+
+        cy.readFile('cypress/fixtures/CQLForTestCaseExecution.txt').should('exist').then((fileContents) => {
+            cy.get(EditMeasurePage.cqlEditorTextBox).type(fileContents)
+        })
+
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.enabled')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).click()
+        
+        //Create Measure Group
+        cy.get(EditMeasurePage.measureGroupsTab).should('be.visible')
+        cy.get(EditMeasurePage.measureGroupsTab).click()
+        
+        cy.get(MeasureGroupPage.initialPopulationSelect).select('ipp')
+        cy.get(MeasureGroupPage.denominatorSelect).select('denom')
+        cy.get(MeasureGroupPage.numeratorSelect).select('num')
+        cy.get(MeasureGroupPage.numeratorExclusionSelect).select('num')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.visible')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.enabled')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).click()
+        
+        //validation successful save message
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('exist')
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('contain.text', 'Population details for this group saved successfully.')
+
+        //TestCasesPage.createTestCase(testCaseTitle, testCaseDescription, testCaseSeries, validTestCaseJson, true)
+
+        //Navigate to Test Cases page and add Test Case details
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+        cy.get(TestCasesPage.newTestCaseButton).should('be.visible')
+        cy.get(TestCasesPage.newTestCaseButton).should('be.enabled')
+        cy.get(TestCasesPage.newTestCaseButton).click()
+        
+        cy.get(TestCasesPage.testCasePopulationList).should('be.visible')
+        
+        cy.get(TestCasesPage.testCaseTitle).should('be.visible')
+        cy.get(TestCasesPage.testCaseTitle).should('be.enabled')
+        cy.get(TestCasesPage.testCaseTitle).type(testCaseTitle, { force: true })
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).type(testCaseDescription)
+        cy.get(TestCasesPage.testCaseSeriesTextBox).type(testCaseSeries).type('{enter}')
+
+        cy.readFile('cypress/fixtures/measureId').should('exist').then((id)=> {
+            cy.intercept('POST', '/api/measures/' + id + '/test-cases').as('testcase')
+        
+            cy.get(TestCasesPage.createTestCaseButton).should('be.visible')
+            cy.get(TestCasesPage.createTestCaseButton).should('be.enabled')
+            cy.get(TestCasesPage.createTestCaseButton).click()
+        
+            //saving testCaseId to file to use later
+            cy.wait('@testcase').then(({response}) => {
+                expect(response.statusCode).to.eq(201)
+                cy.writeFile('cypress/fixtures/testCaseId', response.body.id)
+            })
+
+            cy.get(TestCasesPage.confirmationMsg).should('contains.text', 'Test case created successfully!')
+            
+            cy.get(EditMeasurePage.testCasesTab).click()
+        
+        })
+
+        TestCasesPage.clickEditforCreatedTestCase()
+        cy.get(TestCasesPage.runTestButton).should('be.disabled')
+
+        cy.get(EditMeasurePage.testCasesTab).should('be.visible')
+        cy.get(EditMeasurePage.testCasesTab).click()
+
+        cy.get(TestCasesPage.executeTestCaseButton).should('be.disabled')
+
+    })
+    it('Run Test Case actual results in Population Values table', () => {
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Add CQL
+        cy.get(EditMeasurePage.cqlEditorTab).click()
+        
+        cy.readFile('cypress/fixtures/CQLForTestCaseExecution.txt').should('exist').then((fileContents) => {
+            cy.get(EditMeasurePage.cqlEditorTextBox).type(fileContents)
+        })
+        
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.visible')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).should('be.enabled')
+        cy.get(EditMeasurePage.cqlEditorSaveButton).click()
+                
+        //Create Measure Group
+        cy.get(EditMeasurePage.measureGroupsTab).should('be.visible')
+        cy.get(EditMeasurePage.measureGroupsTab).click()
+
+        //select a group type
+        cy.get(MeasureGroupPage.measureGroupTypeSelect).should('exist')
+        cy.get(MeasureGroupPage.measureGroupTypeSelect).should('be.visible')
+        cy.get(MeasureGroupPage.measureGroupTypeSelect).click()
+        cy.get(MeasureGroupPage.measureGroupTypeCheckbox).each(($ele) => {
+            if ($ele.text() == "Process") {
+                cy.wrap($ele).should('exist')
+                cy.wrap($ele).focus()
+                cy.wrap($ele).click()
+            }
+        })
+        cy.get(MeasureGroupPage.measureGroupTypeDropdownBtn).should('exist').invoke('click')        
+                
+        cy.get(MeasureGroupPage.initialPopulationSelect).select('ipp')
+        cy.get(MeasureGroupPage.denominatorSelect).select('denom')
+        cy.get(MeasureGroupPage.numeratorSelect).select('num')
+        cy.get(MeasureGroupPage.numeratorExclusionSelect).select('num')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.visible')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).should('be.enabled')
+        cy.get(MeasureGroupPage.saveMeasureGroupDetails).click()
+                
+        //validation successful save message
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('exist')
+        cy.get(MeasureGroupPage.successfulSaveMeasureGroupMsg).should('contain.text', 'Population details for this group updated successfully.')
+        
+        TestCasesPage.createTestCase(testCaseTitle, testCaseDescription, testCaseSeries, validTestCaseJson, true)
+        TestCasesPage.clickEditforCreatedTestCase()
+
+        cy.get(TestCasesPage.detailsTab).click()
+
+        cy.get(TestCasesPage.testCaseIPPCheckBox).check().should('be.checked')
+        cy.get(TestCasesPage.testCaseDENOMCheckBox).check().should('be.checked')
+        cy.get(TestCasesPage.testCaseDENEXCheckBox).check().should('be.checked')
+
+
+        cy.get(TestCasesPage.createTestCaseButton).should('be.visible')
+        cy.get(TestCasesPage.createTestCaseButton).should('be.enabled')
+        cy.get(TestCasesPage.createTestCaseButton).click()
+
+        cy.get(TestCasesPage.runTestButton).should('be.visible')
+        cy.get(TestCasesPage.runTestButton).should('be.enabled')
+        cy.get(TestCasesPage.runTestButton).click()
+
+        cy.get(TestCasesPage.ippActualCheckBox).should('be.checked')
+        cy.get(TestCasesPage.denomActualCheckBox).should('be.checked')
+        cy.get(TestCasesPage.denomExclusionActualCheckBox).should('be.checked')
+
+
+    })
+})
