+++ cypress/integration/WebInterface/Test Cases/Validations/TestCaseValidations.spec.ts
@@ -1,0 +1,128 @@
+import {OktaLogin} from "../../../../Shared/OktaLogin"
+import {CreateMeasurePage} from "../../../../Shared/CreateMeasurePage"
+import {MeasuresPage} from "../../../../Shared/MeasuresPage"
+import {TestCasesPage} from "../../../../Shared/TestCasesPage"
+import {EditMeasurePage} from "../../../../Shared/EditMeasurePage"
+import {TestCaseJson} from "../../../../Shared/TestCaseJson"
+import {Utilities} from "../../../../Shared/Utilities"
+
+let measureName = 'TestMeasure' + Date.now()
+let CqlLibraryName = 'TestLibrary' + Date.now()
+let testCaseTitle = 'test case title'
+let testCaseDescription = 'DENOMFail' + Date.now()
+let validTestCaseJson = TestCaseJson.TestCaseJson_Valid
+let testCaseSeries = 'SBTestSeries'
+let twoFiftyTwoCharacters = 'abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijklmnopqr'
+
+describe('Test Case Validations', () => {
+
+    before('Create Measure', () => {
+        //Create New Measure
+        CreateMeasurePage.CreateQICoreMeasureAPI(measureName, CqlLibraryName)
+    })
+
+    beforeEach('Login', () => {
+        OktaLogin.Login()
+    })
+    afterEach('Logout', () => {
+        OktaLogin.Logout()
+
+    })
+
+    after('Clean up', () => {
+
+        Utilities.deleteMeasure(measureName, CqlLibraryName)
+
+    })
+
+    it('Create Test Case: Description more than 250 characters', () => {
+
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Navigate to Test Cases page and add Test Case details
+        cy.get(EditMeasurePage.testCasesTab).click()
+        cy.get(TestCasesPage.newTestCaseButton).click()
+        cy.get(TestCasesPage.detailsTab).click()
+        cy.get(TestCasesPage.testCaseTitle).type(testCaseTitle)
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).type(twoFiftyTwoCharacters)
+        cy.get(TestCasesPage.createTestCaseButton).should('be.disabled')
+    })
+
+    it('Edit Test Case: Description more than 250 characters', () => {
+
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        TestCasesPage.createTestCase(testCaseTitle, testCaseDescription, testCaseSeries, validTestCaseJson, true)
+
+        //Click on Edit for Test Case
+        TestCasesPage.clickEditforCreatedTestCase()
+
+        cy.get(TestCasesPage.detailsTab).click()
+
+        //Update Test Case Description with more than 250 characters
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).clear()
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).type(twoFiftyTwoCharacters, {delay: 0})
+        cy.get(TestCasesPage.cuTestCaseButton).should('be.disabled')
+    })
+
+    it('Create Test Case: Title more than 250 characters', () => {
+
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        //Navigate to Test Cases page and add Test Case details
+        cy.get(EditMeasurePage.testCasesTab).click()
+        cy.get(TestCasesPage.newTestCaseButton).should('be.visible')
+        cy.get(TestCasesPage.newTestCaseButton).should('be.enabled')
+        cy.get(TestCasesPage.newTestCaseButton).click()
+
+        cy.get(TestCasesPage.detailsTab).click()
+
+        cy.get(TestCasesPage.testCasePopulationList).should('be.visible')
+
+        cy.wait(2000)
+
+        cy.get(TestCasesPage.testCaseTitle).should('be.visible')
+        cy.get(TestCasesPage.testCaseTitle).should('be.enabled')
+
+        cy.get(TestCasesPage.testCaseTitle).type(twoFiftyTwoCharacters)
+        cy.get(TestCasesPage.testCaseDescriptionTextBox).type(testCaseDescription)
+        cy.get(TestCasesPage.createTestCaseButton).should('be.disabled')
+        cy.get(TestCasesPage.testCaseTitleInlineError).contains('Test Case Title cannot be more ' +
+            'than 250 characters.')
+
+    })
+
+    it('Edit Test Case: Title more than 250 characters', () => {
+
+        //Click on Edit Measure
+        MeasuresPage.clickEditforCreatedMeasure()
+
+        TestCasesPage.createTestCase(testCaseTitle, testCaseDescription, testCaseSeries, validTestCaseJson, true)
+
+        //Click on Edit for Test Case
+        TestCasesPage.clickEditforCreatedTestCase()
+
+        cy.get(TestCasesPage.detailsTab).click()
+
+        cy.get(TestCasesPage.testCasePopulationList).should('be.visible')
+
+        cy.wait(2000)
+
+        cy.get(TestCasesPage.testCaseTitle).should('exist')
+        cy.get(TestCasesPage.testCaseTitle).should('be.visible')
+        cy.get(TestCasesPage.testCaseTitle).should('be.enabled')
+        cy.get(TestCasesPage.testCaseTitle).focus().clear()
+        cy.get(TestCasesPage.testCaseTitle).invoke('val', '')
+        cy.get(TestCasesPage.testCaseTitle).type('{selectall}{backspace}{selectall}{backspace}')
+
+        //Update Test Case Description with more than 250 characters
+        cy.get(TestCasesPage.testCaseTitle).type(twoFiftyTwoCharacters, {delay: 0})
+        cy.get(TestCasesPage.testCaseSeriesTextBox).click()
+        cy.get(TestCasesPage.cuTestCaseButton).should('be.disabled')
+        cy.get(TestCasesPage.testCaseTitleInlineError).contains('Test Case Title cannot be more ' +
+            'than 250 characters.')
+    })
+})
